Set failure error on POST and fix minor issues
